Fix hassio warnings and false True (positive) for DisableLinkage feature (#394)
* Fix hassio warnings

setup without awaiting async_forward_entry_setup, which can cause the setup lock to be released before the setup is done. This will stop working in Home Assistant 2025.1

Detected blocking call to load_default_certs

Signed-off-by: Jan Dušek <jan.dusek4@gmail.com>

* Changing query for DisableLinkage

Unify the way it is done elsewhere to prevent excesive false positive retun of True when DisableLinkage is actually available and returns error from camera. After thsi fix, it is queried twice and then the integration knows it is unavailable.

Signed-off-by: Jan Dušek <jan.dusek4@gmail.com>

---------

Signed-off-by: Jan Dušek <jan.dusek4@gmail.com>